REFACTOR: extract ClockWithTooltip component from App
- App에서 ClockWithTooltip 컴포넌트 분리
