+++ src/components/index.ts
@@ -1,4 +1,5 @@
 import Clock from "./Clock/Clock";
 import Tooltip from "./Tooltip/Tooltip";
+import ClockWithTooltip from "./ClockWithTooltip/ClockWithTooltip";
 
-export { Clock, Tooltip };
+export { Clock, Tooltip, ClockWithTooltip };
